🚚 Renamed folders in i18n (World -> Pages)
