+++ src/i18n/en-US/Pages/World/console.ts
