+++ src/i18n/en-US/Pages/World/property.ts
