+++ src/i18n/en-US/index.ts
@@ -1,22 +1,22 @@
 import { MessageSchema } from "src/boot/i18n";
-import { enUSHome } from "src/i18n/en-US/World/home";
-import { enUSproperty } from "src/i18n/en-US/World/property";
-import { enUSConsole } from "./World/console";
-import { enUSPlayer } from "./World/player";
-import { enUSSystemSetting } from "./Other/systemsetting";
-import { enUSBugReport } from "./Other/bugReport";
-import { enAdditionalContents } from "./World/additionalContents";
+import { enUSHome } from "src/i18n/en-US/Pages/World/home";
+import { enUSproperty } from "src/i18n/en-US/Pages/World/property";
+import { enUSConsole } from "./Pages/World/console";
+import { enUSPlayer } from "./Pages/World/player";
+import { enUSSystemSetting } from "./Pages/systemsetting";
+import { enUSBugReport } from "./Pages/bugReport";
+import { enAdditionalContents } from "./Pages/World/additionalContents";
 import { enUSError } from "./Other/error";
-import { enUSProgress } from "./World/progress";
-import { enUSMainLayout } from "./World/mainLayout";
+import { enUSProgress } from "./Pages/World/progress";
+import { enUSMainLayout } from "./Pages/mainLayout";
 import { enUSIcon } from "./Dialogs/icon";
-import { enUSShareWorld } from "./World/shareWorld";
+import { enUSShareWorld } from "./Pages/World/shareWorld";
 import { enWelcome } from "./Dialogs/welcome";
 import { enAutoShutdown } from "./Dialogs/autoshutdown";
 import { enUSGeneral } from "./Other/general";
 import { enUSEulaDialog } from "./Dialogs/eulaDialog";
 import { enUSOwner } from "./Dialogs/owner";
-import { enOthers } from "./World/others";
+import { enOthers } from "./Pages/World/others";
 import { enTabs } from "./Other/tabs";
 
 export const enUS: MessageSchema = {
