+++ src/i18n/ja/Pages/World/additionalContents.ts
