+++ src/i18n/ja/Pages/World/console.ts
