+++ src/i18n/ja/Pages/World/home.ts
