+++ src/i18n/ja/Pages/World/others.ts
