+++ src/i18n/ja/Pages/World/player.ts
