+++ src/i18n/ja/Pages/World/progress.ts
