+++ src/i18n/ja/Pages/World/property.ts
