+++ src/i18n/ja/Pages/World/shareWorld.ts
