+++ src/i18n/ja/Pages/bugReport.ts
