+++ src/i18n/ja/Pages/mainLayout.ts
